pyproject.toml
[tool.poetry]
name = "lab"
version = "0.1.0"
description = ""
authors = ["meirelesgc <geu_costa@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.2"}
psycopg = {extras = ["binary"], version = "^3.2.1"}
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.14"
pypdf = "^4.3.1"
openai = "^1.43.0"
tiktoken = "^0.7.0"
celery = {extras = ["rabitmq"], version = "^5.4.0"}
flower = "^2.0.1"
spacy = "^3.7.6"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.6.3"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'uvicorn lab:app --reload'
test = 'pytest -s -x --cov=lab -vv'
flower = 'celery -A lab.celery flower'
worker = 'celery -A lab.routers.documents.celery worker --loglevel=INFO --concurrency=1'